:recycle: refactor: event로 contestId를 받아오지 않고 현재 진행 중인 콘테스트를 모두 update  하도록
