+++ src/main/java/com/mulmeong/contest/application/ContestService.java
@@ -2,10 +2,7 @@ package com.mulmeong.contest.application;
 
 import com.mulmeong.contest.dto.in.ContestRequestDto;
 import com.mulmeong.contest.dto.in.PostRequestDto;
-import com.mulmeong.contest.dto.in.PostVoteRenewRequestDto;
 import com.mulmeong.contest.dto.in.PostVoteRequestDto;
-import com.mulmeong.event.contest.ContestPostUpdateEvent;
-import com.mulmeong.event.contest.ContestVoteRenewEvent;
 
 public interface ContestService {
 
@@ -15,6 +12,6 @@ public interface ContestService {
 
     void vote(PostVoteRequestDto dto, String memberUuid);
 
-    void voteRenew(ContestVoteRenewEvent message);
+    void voteRenew();
 
 }
