+++ src/main/java/com/mulmeong/contest/application/KafkaConsumer.java
@@ -1,6 +1,5 @@
 package com.mulmeong.contest.application;
 
-import com.mulmeong.event.contest.ContestVoteRenewEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -15,9 +14,8 @@ public class KafkaConsumer {
 
     @KafkaListener(topics = "${event.contest.pub.topics.contest-vote-renew.name}",
             containerFactory = "contestVoteRenewListener")
-    public void updateContestPost(ContestVoteRenewEvent message) {
-        log.info("vote message: {}", message.toString());
-        contestService.voteRenew(message);
+    public void updateContestPost() {
+        contestService.voteRenew();
     }
