+++ src/main/java/com/mulmeong/contest/infrastructure/ContestRepository.java
@@ -4,7 +4,11 @@ import com.mulmeong.contest.entity.Contest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDate;
+import java.util.List;
 
 public interface ContestRepository extends JpaRepository<Contest, Long> {
     Iterable<Contest> findByEndDate(LocalDate localDate);
+
+    List<Contest> findByStartDateLessThanEqualAndEndDateGreaterThanEqual(LocalDate startDate, LocalDate endDate);
+
 }
