--- src/main/java/com/mulmeong/event/contest/ContestVoteRenewEvent.java
@@ -1,14 +1,0 @@
-package com.mulmeong.event.contest;
-
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Data
-@Getter
-@NoArgsConstructor
-public class ContestVoteRenewEvent {
-
-    private Long ContestId;
-
-}
